requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
seaborn
google-generativeai
PyGithub
openpyxl
